+++ crates/megane-core/src/lib.rs
@@ -7,6 +7,7 @@ pub mod dcd;
 pub mod gro;
 pub mod lammps_data;
 pub mod lammpstrj;
+pub mod magres;
 pub mod mmcif;
 pub mod mol;
 pub mod mol2;
@@ -19,6 +20,5 @@ pub mod trajectory;
 pub mod vasp;
 pub mod xtc;
 pub mod xyz;
-
 pub use parser::ParsedStructure;
 pub use trajectory::TrajectoryData;