+++ python/megane/parsers/__init__.py
@@ -3,6 +3,7 @@ from megane.parsers.dcd import load_dcd
 from megane.parsers.gro import load_gro
 from megane.parsers.lammps_data import load_lammps_data
 from megane.parsers.lammpstrj import load_lammpstrj, load_lammpstrj_structure
+from megane.parsers.magres import load_magres
 from megane.parsers.mol import load_mol, load_sdf
 from megane.parsers.mol2 import load_mol2
 from megane.parsers.netcdf import load_netcdf
@@ -19,6 +20,7 @@ __all__ = [
     "load_lammps_data",
     "load_lammpstrj",
     "load_lammpstrj_structure",
+    "load_magres",
     "load_mol",
     "load_mol2",
     "load_netcdf",